✨ AtomIOInternalResource allows for the simplification of signatures for withdraw and deposit
