+++ packages/atom.io/realtime-server/src/realtime-server-stores/server-room-external-actions.ts
@@ -1,5 +1,5 @@
 import * as AtomIO from "atom.io"
-import { editRelations, editRelationsInStore, type Loadable } from "atom.io/data"
+import { editRelationsInStore, type Loadable } from "atom.io/data"
 import type { UserInRoomMeta } from "atom.io/realtime"
 import { roomIndex, usersInRooms } from "atom.io/realtime"
